Fix non-working /tf publication
When using the client as such:
rosrun rc_hand_eye_calibration_client rc_hand_eye_calibration_client_node  _host:=10.0.2.40 _calibration_publication_period:=1 _calibration_request_period:=0

First, the calibration request is done. During the answer on line 195, start()
is called on the not yet "created" (with NodeHandle::createSteadyTimer()),
but default-constructed calib_publish_timer (the timer for publication),

The default-constructed timer starts with an empty callback and stores that
it is started. Therefore it doesn't react to further start() calls

Solution:
- Create the timer before starting it.
- To make future problems unlikely, check validity of SteadyTimer before
  using start()