rc_visard_driver: fix race condition when changing exposure mode from auto to manual
If camera_exp_auto (ExposureAuto) is changed, immediately set new value in
dynamic reconfigure callback so that the updated ExposureTime and Gain can
be read and published again.
If we do this in the grabbing thread we have a race condition and another
parameter update in the meantime could be lost because we need to publish/overwrite
the config with the updated time and gain values.